third_party/blink/web_tests/virtual/deprecate-unload/external/wpt/fetch/api/redirect/redirect-keepalive.any-expected.txt

This is a testharness.js-based test.
[PASS] [keepalive][new window][unload] same-origin redirect; setting up
[PASS] [keepalive][new window][unload] same-origin redirect + preflight; setting up
[PASS] [keepalive][new window][unload] cross-origin redirect; setting up
[PASS] [keepalive][new window][unload] cross-origin redirect + preflight; setting up
[PASS] [keepalive][new window][unload] redirect to file URL; setting up
[PASS] [keepalive][new window][unload] redirect to data URL; setting up
[FAIL] [keepalive][new window][unload] same-origin redirect
  assert_equals: token should be on (stashed). expected (string) "on" but got (object) null
[FAIL] [keepalive][new window][unload] same-origin redirect + preflight
  assert_equals: token should be on (stashed). expected (string) "on" but got (object) null
[FAIL] [keepalive][new window][unload] cross-origin redirect
  assert_equals: token should be on (stashed). expected (string) "on" but got (object) null
[FAIL] [keepalive][new window][unload] cross-origin redirect + preflight
  assert_equals: token should be on (stashed). expected (string) "on" but got (object) null
[PASS] [keepalive][new window][unload] redirect to file URL
[PASS] [keepalive][new window][unload] redirect to data URL
Harness: the test ran to completion.
